+++ globals.go
@@ -22,6 +22,9 @@ func GetOptions() []Option { //nolint:unused
 	o := *options.Load()
 	opts := make([]Option, 0)
 
+	if o.allowNonAtomicReplace != optionDefaults.allowNonAtomicReplace {
+		opts = append(opts, WithAllowNonAtomicReplace(o.allowNonAtomicReplace))
+	}
 	if o.atomically != optionDefaults.atomically {
 		opts = append(opts, WithAtomicity(o.atomically))
 	}
@@ -68,6 +71,8 @@ func buildOptions(opts ...Option) Options { //nolint:unused
 
 func (o Options) String() string { //nolint:unused
 	var b strings.Builder
+
+	fmt.Fprintf(&b, "allowNonAtomicReplace:      %v\n", o.allowNonAtomicReplace)
 	fmt.Fprintf(&b, "atomically:      %v\n", o.atomically)
 	fmt.Fprintf(&b, "defaultFileMode: 0o%03o (%v)\n", o.defaultFileMode, o.defaultFileMode)
 	fmt.Fprintf(&b, "fileMode:        0o%03o (%v)\n", o.fileMode, o.fileMode)