+++ writefile.go
@@ -28,6 +28,11 @@ import (
 //
 // Additional option arguments can be used to change the default configuration
 // for the target file.
+//
+// On Plan 9, atomic creation of a new file is supported, but atomic replacement
+// of an existing file is not. If the destination exists, WriteFile returns an
+// error matching errors.ErrUnsupported and leaves the destination unchanged.
+// To work around this issue, use the WithAllowNonAtomicReplace option.
 func WriteFile(name string, data []byte, perm os.FileMode, opts ...Option) error {
 	return WriteReader(name, bytes.NewReader(data), perm, opts...)
 }